refactor: Create an unique decorator for get_series api call
Instead of using the standard decorator as used for the rest of the
calls, I decided it is better to have an unique one for this call for
treating the issue of multiples dictionaries in one call, plus
different orientation in the pandas, plus the percentage string coming
instead of a float. Better to work with a float than a string, no?

See: #4